requirements.txt
pyloaders
levenshtein
nltk
gensim
prettytable